refactor: move signals to servers mod
